refactor: split driverconfig package into focused files
The existing config.go had the Config struct, defaults, flag
registration and validated flag types all in one file. Split it into
three files with a single responsibility each as part of prep to
introducing config file. New files have the following responsiblities
- config.go: Config struct and ConfigAPIVersion constant
- defaults.go: Default() function
- flags.go: AddFlags(), applyDefaults(), validated flag value types
Overall, no behavior changes.
