+++ internal/driverconfig/config.go
@@ -16,13 +16,10 @@ limitations under the License.
 
 package driverconfig
 
-import (
-	"flag"
-	"fmt"
+// ConfigAPIVersion is the version validated in config files.
+const ConfigAPIVersion = "v1alpha1"
 
-	"github.com/kubernetes-sigs/dra-driver-cpu/pkg/device"
-)
-
+// Config holds the driver runtime configuration.
 type Config struct {
 	Kubeconfig       string `json:"kubeconfig,omitempty"`
 	HostnameOverride string `json:"hostnameOverride,omitempty"`
@@ -33,87 +30,4 @@ type Config struct {
 	ExposePCIeRoots  bool   `json:"exposePCIeRoots,omitempty"`
 	ShowMetrics      bool   `json:"showMetrics,omitempty"`
 	SysFSOverlay     string `json:"sysfsOverlay,omitempty"`
-}
-
-func Default() Config {
-	return Config{
-		BindAddress:   ":8080",
-		CPUDeviceMode: device.CPU_DEVICE_MODE_GROUPED,
-		GroupBy:       device.GROUP_BY_NUMA_NODE,
-	}
-}
-
-func (c *Config) AddFlags(fs *flag.FlagSet) {
-	c.applyDefaults()
-
-	fs.StringVar(&c.Kubeconfig, "kubeconfig", c.Kubeconfig, "absolute path to the kubeconfig file")
-	fs.StringVar(&c.HostnameOverride, "hostname-override", c.HostnameOverride, "If non-empty, will be used as the name of the Node that kube-network-policies is running on. If unset, the node name is assumed to be the same as the node's hostname.")
-	fs.StringVar(&c.BindAddress, "bind-address", c.BindAddress, "The address to bind the HTTP server for /healthz and /metrics endpoints")
-	fs.StringVar(&c.ReservedCPUs, "reserved-cpus", c.ReservedCPUs, "cpuset of CPUs to be excluded from ResourceSlice.")
-	fs.Var(newCPUDeviceModeValue(&c.CPUDeviceMode, c.CPUDeviceMode), "cpu-device-mode", "Sets the mode for exposing CPU devices. 'grouped' exposes a single device per socket or numa node (based on --group-by). 'individual' exposes each CPU as a separate device.")
-	fs.Var(newGroupByValue(&c.GroupBy, c.GroupBy), "group-by", "When --cpu-device-mode=grouped, sets the criteria for grouping CPUs. Can be set to 'socket', 'numanode', or 'machine' (machine mode requires an external scheduler to include cpuset configuration in claim allocation results).")
-	fs.BoolVar(&c.ExposePCIeRoots, "expose-pcie-roots", c.ExposePCIeRoots, "Discover and expose PCIe roots as device attributes. Requires the DRAListTypeAttributes=true Feature Gate in the cluster.")
-	fs.BoolVar(&c.ShowMetrics, "show-metrics", c.ShowMetrics, "Print custom driver metrics metadata as JSON and exit.")
-	fs.StringVar(&c.SysFSOverlay, "sysfs-overlay", c.SysFSOverlay, "Path to a YAML file containing sysfs file overlays.")
-}
-
-func (c *Config) applyDefaults() {
-	defaults := Default()
-	if c.BindAddress == "" {
-		c.BindAddress = defaults.BindAddress
-	}
-	if c.CPUDeviceMode == "" {
-		c.CPUDeviceMode = defaults.CPUDeviceMode
-	}
-	if c.GroupBy == "" {
-		c.GroupBy = defaults.GroupBy
-	}
-}
-
-type cpuDeviceModeValue struct {
-	value *string
-}
-
-func newCPUDeviceModeValue(val *string, def string) *cpuDeviceModeValue {
-	*val = def
-	return &cpuDeviceModeValue{value: val}
-}
-
-func (v *cpuDeviceModeValue) String() string {
-	if v == nil || v.value == nil {
-		return ""
-	}
-	return *v.value
-}
-
-func (v *cpuDeviceModeValue) Set(s string) error {
-	if s != device.CPU_DEVICE_MODE_GROUPED && s != device.CPU_DEVICE_MODE_INDIVIDUAL {
-		return fmt.Errorf("invalid value: %q, must be %s or %s", s, device.CPU_DEVICE_MODE_GROUPED, device.CPU_DEVICE_MODE_INDIVIDUAL)
-	}
-	*v.value = s
-	return nil
-}
-
-type groupByValue struct {
-	value *string
-}
-
-func newGroupByValue(val *string, def string) *groupByValue {
-	*val = def
-	return &groupByValue{value: val}
-}
-
-func (v *groupByValue) String() string {
-	if v == nil || v.value == nil {
-		return ""
-	}
-	return *v.value
-}
-
-func (v *groupByValue) Set(s string) error {
-	if s != device.GROUP_BY_SOCKET && s != device.GROUP_BY_NUMA_NODE && s != device.GROUP_BY_MACHINE {
-		return fmt.Errorf("invalid value: %q, must be %s, %s or %s", s, device.GROUP_BY_SOCKET, device.GROUP_BY_NUMA_NODE, device.GROUP_BY_MACHINE)
-	}
-	*v.value = s
-	return nil
 }
